AP_Vehicle: change namespace of MultiCopter and FixedWing params
this stops the libraries knowing anything about AP_Vehicle
