+++ libraries/AP_Vehicle/AP_MultiCopter.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <AP_Param/AP_Param.h>
+
+/*
+  common parameters for multicopters
+*/
+struct AP_MultiCopter {
+    AP_Int16 angle_max;
+};
